Build FvwmPrompt and remove FvwmConsole.
Build FvwmPrompt and remove use of FvwmConsole. With FvwmPrompt
dropping the use if ishell, all the go depends are now present in
Debian and FvwmPrompt can be built without use of the vendored
libraries, and will now be used instead of FvwmConsole.

This requires some "hacks" in debian/rules to use the debhelper
tools to build FvwmPrompt independent of meson.